% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_api_data.R
\name{clean_api_data}
\alias{clean_api_data}
\title{Cleans API data}
\usage{
clean_api_data(list1)
}
\arguments{
\item{list1}{object as a nested list containing annual voting data}
}
\value{
returns a list of elements where each element is a data frame of annual voting by party
}
\description{
takes list input from the get_riksdagen_api_data() and prepares dataset for Shiny App usage
}
