% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riksdagenapi.R
\name{get_riksdagen_api_data}
\alias{get_riksdagen_api_data}
\title{API Data pull from Swedish Riksdagen}
\usage{
get_riksdagen_api_data()
}
\value{
returns a list of elements where each element is a data frame of annual voting by member
}
\description{
connects to the Riksdagen API and pulls the member wise voting data from 2009 to 2020
}
