% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny_app.R
\name{shinyFunction}
\alias{shinyFunction}
\title{Shiny App}
\usage{
shinyFunction(dataframe_1)
}
\arguments{
\item{dataframe_1}{as a list if data frames from clean_api_data function}
}
\description{
creates visualization tool for users
}
